+++ src/main/java/chess/domain/board/dto/BoardDto.java
@@ -1,4 +1,4 @@
-package chess.web.dto;
+package chess.domain.board.dto;
 
 import chess.domain.board.Board;
 import chess.domain.game.ChessGame;
