+++ src/main/java/chess/web/controller/ChessController.java
@@ -5,10 +5,10 @@ import static chess.web.view.RenderView.renderHtml;
 import chess.domain.game.ChessGame;
 import chess.domain.game.dao.ChessGameDao;
 import chess.service.ChessService;
-import chess.web.dto.BoardDto;
-import chess.web.dto.ChessGameDto;
-import chess.web.dto.GameNameDto;
-import chess.web.dto.MoveCommandDto;
+import chess.domain.board.dto.BoardDto;
+import chess.domain.game.dto.ChessGameDto;
+import chess.domain.game.dto.GameNameDto;
+import chess.domain.command.dto.MoveCommandDto;
 import java.util.HashMap;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
