+++ src/test/java/chess/domain/command/dao/MoveCommandDaoTest.java
@@ -1,0 +1,36 @@
+package chess.domain.command.dao;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.TestPropertySource;
+
+@JdbcTest
+@AutoConfigureTestDatabase(replace = Replace.NONE)
+@TestPropertySource("classpath:application-test.properties")
+class MoveCommandDaoTest {
+
+    private MoveCommandDao moveCommandDao;
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    @BeforeEach
+    void setUp() {
+        moveCommandDao = new MoveCommandDao(jdbcTemplate);
+    }
+
+    @Test
+    void addMoveCommand() {
+
+    }
+
+    @Test
+    void findCommandsByGameId() {
+
+    }
+}
