+++ src/test/java/chess/domain/game/dao/ChessGameDaoTest.java
@@ -3,7 +3,7 @@ package chess.domain.game.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.game.ChessGame;
-import chess.web.dto.ChessGameDto;
+import chess.domain.game.dto.ChessGameDto;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
